chore[output-mapping]: Merge AbstractWorkspaceTableStrategy with SqlWorkspaceTableStrategy (only implementation)
